Fix frequency
In skyrim, loading screens with correct conditions are prioritized over loading screens with no conditions.
In order to get 30% frequency, the probability of the conditions of all loading screens being wrong must be 70%.
Then in 70% of the cases, a condition-less loading screen will be used and in 30% of the cases a loading screen of the mod will be used.